[storage] Introduce storage_units and migrate Fxfs
Introduce the storage_units crate (//src/storage/lib/units) to represent
power-of-two block sizes and system page sizes as zero-cost typed
abstractions (GenericBlockSize, ValueBlockSize, and PAGE_SIZE).

Block sizes are guaranteed to be powers of two, replacing expensive
64-bit integer division and modulo operations with single-cycle bit
shifts and masks. Standard arithmetic, comparison, and formatting traits
are implemented, along with overflow-safe alignment helpers (align_up,
align_down, align_up_to_blocks, and is_aligned). On Fuchsia, PAGE_SIZE
caches the system page mask using an atomic to avoid repeated
zx::system_get_page_size() calls.

Migrate Fxfs to use BlockSize and PAGE_SIZE across the filesystem.

TAG: agy
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1807218
Commit-Queue: Chris Drouillard <cdrllrd@google.com>

+++ src/storage/f2fs_reader/migrate/src/integration_test.rs
@@ -36,7 +36,7 @@ fn create_device_with_image_at_offset(path: &str, offset: u64) -> Arc<VmoBackedS
     // so we'll resize dynamically.
     let mut current_size = std::cmp::max(
         offset + 1024 * 1024, // Start with at least 1MB past offset
-        SuperBlockInstance::B.first_extent().end + FXFS_BLOCK_SIZE as u64,
+        SuperBlockInstance::B.first_extent().end + FXFS_BLOCK_SIZE,
     );
     let vmo = zx::Vmo::create_with_opts(zx::VmoOptions::RESIZABLE, current_size)
         .expect("failed to create vmo");
@@ -64,7 +64,7 @@ fn create_device_with_image_at_offset(path: &str, offset: u64) -> Arc<VmoBackedS
     }
 
     // Ensure the VMO is at least the size required for Fxfs superblocks.
-    let min_size = SuperBlockInstance::B.first_extent().end + FXFS_BLOCK_SIZE as u64;
+    let min_size = SuperBlockInstance::B.first_extent().end + FXFS_BLOCK_SIZE;
     if current_offset < min_size {
         if current_size < min_size {
             vmo.set_size(min_size).expect("failed to resize vmo");
@@ -214,7 +214,7 @@ async fn test_fxfs_migration_no_keys() {
 #[fuchsia::test]
 async fn test_fxfs_migration_with_offset() {
     // Place f2fs image between Fxfs superblocks.
-    let offset = SuperBlockInstance::A.first_extent().end + FXFS_BLOCK_SIZE as u64;
+    let offset = SuperBlockInstance::A.first_extent().end + FXFS_BLOCK_SIZE;
     Box::pin(test_fxfs_migration_at_offset(offset)).await;
 }
